Moved user endpoints to /api/user/
