+++ api/endpoints/data.py
@@ -16,7 +16,7 @@ from api.helpers import calculate_time_args, is_admin, data_cleaner
 from api.user import get_user_level
 
 
-api_bp = Blueprint('api_bp', __name__, static_url_path='')
+data_api_bp = Blueprint('data_api_bp', __name__, static_url_path='')
 
 # decorator to limit certain pages to a specific user level
 def required_user_level(level_config_key):
@@ -60,12 +60,12 @@ def required_user_level(level_config_key):
 ##
 ## Example:
 ## http://127.0.0.1:5000/api/
-@api_bp.route('/', methods=["GET"])
+@data_api_bp.route('/', methods=["GET"])
 def health():
     return make_response(jsonify( dt.datetime.now(dt.timezone.utc) ), 200)
 
 ## Helper function needed for accessing raw list of all meters in other blueprint
-@api_bp.route('/meters')
+@data_api_bp.route('/meters')
 @required_user_level("USER_LEVEL_VIEW_DASHBOARD")
 def meters():
     statement = db.select(models.Meter)
@@ -87,7 +87,7 @@ def meters():
     return make_response(jsonify(out), 200)
 
 ## Health check cache meta
-@api_bp.route('/hc_meta')
+@data_api_bp.route('/hc_meta')
 @required_user_level("USER_LEVEL_VIEW_HEALTHCHECK")
 def hc_meta():
     hc_meta = db.session.execute(db.select(models.CacheMeta).where(models.CacheMeta.meta_type == "health_check")).scalar_one_or_none()
@@ -162,7 +162,7 @@ def hc_meta():
 ##
 ## Example:
 ## http://127.0.0.1:5000/api/summary
-@api_bp.route('/summary')
+@data_api_bp.route('/summary')
 @required_user_level("USER_LEVEL_VIEW_DASHBOARD")
 def summary():
     to_time = request.args.get("to_time")
@@ -217,7 +217,7 @@ def summary():
 ## http://127.0.0.1:5000/api/meter_obs?id=AP001_L01_M2
 ## http://127.0.0.1:5000/api/meter_obs?id=AP001_L01_M2;AP080_L01_M5
 ## http://127.0.0.1:5000/api/meter_obs?id=WTHR_0
-@api_bp.route('/meter_obs')
+@data_api_bp.route('/meter_obs')
 @required_user_level("USER_LEVEL_VIEW_DASHBOARD")
 def meter_obs():
     try:
@@ -292,7 +292,7 @@ def meter_obs():
 ##
 ## Example:
 ## http://127.0.0.1:5000/api/meter_health?id=AP001_L01_M2&date_range=30
-@api_bp.route('/meter_health')
+@data_api_bp.route('/meter_health')
 @required_user_level("USER_LEVEL_VIEW_HEALTHCHECK")
 def meter_health():
     # The frontend should read the headers sent with this response and send another request later to retrieve the latest version if 'X-Cache-State'=stale
@@ -382,7 +382,7 @@ def meter_health():
 ##     },
 ##     ...
 ## }
-@api_bp.route('/meter_hierarchy')
+@data_api_bp.route('/meter_hierarchy')
 @required_user_level("USER_LEVEL_VIEW_DASHBOARD")
 def meter_hierarchy():
     buildings = db.session.execute(db.select(models.Building)).scalars().all()
@@ -436,7 +436,7 @@ def meter_hierarchy():
 ##
 ## Example:
 ## http://127.0.0.1:5000/api/health_score
-@api_bp.route('/health_score')
+@data_api_bp.route('/health_score')
 @required_user_level("USER_LEVEL_VIEW_HEALTHCHECK")
 def health_score():
     to_time = request.args.get("to_time")
@@ -448,7 +448,7 @@ def health_score():
     return make_response(jsonify(data), 200)
 
 ## Returns the contents of data/meta/offline_data.json
-@api_bp.route('/offline_meta')
+@data_api_bp.route('/offline_meta')
 @required_user_level("USER_LEVEL_VIEW_DASHBOARD")
 def offline_meta():
     if not os.path.exists(offline_meta_file) or not offlineMode:
@@ -458,7 +458,7 @@ def offline_meta():
         data = json.load(f)
     return make_response(jsonify(data), 200)
 
-@api_bp.route("/mazemap_polygons")
+@data_api_bp.route("/mazemap_polygons")
 @required_user_level("USER_LEVEL_VIEW_DASHBOARD")
 def mazemap_polygons():
     if not os.path.exists(mazemap_polygons_file):
@@ -468,7 +468,7 @@ def mazemap_polygons():
         data = json.load(f)
     return make_response(jsonify(data), 200)
 
-@api_bp.route('/regeneratecache', methods=["GET"])
+@data_api_bp.route('/regeneratecache', methods=["GET"])
 @required_user_level("USER_LEVEL_ADMIN")
 def regenerate_cache():
     start_time = time.time()
@@ -478,7 +478,7 @@ def regenerate_cache():
     print(f"Cache regeneratation took {total_time} seconds")
     return make_response(str(total_time), 200)
 
-@api_bp.route('/populate_database')
+@data_api_bp.route('/populate_database')
 @required_user_level("USER_LEVEL_ADMIN")
 def populate_database():
     result = initial_database_population()
