+++ dashboard/main.py
@@ -1,9 +1,7 @@
 from flask import render_template, send_file, request, Blueprint, make_response, current_app, redirect, Response, jsonify
 
 from functools import wraps
-from markupsafe import escape
 
-import api.api as api_bp
 import api.user as user
 import dashboard.context as context
 
@@ -51,108 +49,6 @@ def required_user_level(level_config_key):
             return noaccess()
         return wrapper
     return decorator
-
-def setCookies(email: str, sessionID: str) -> Response:
-    resp = make_response(render_template('settings.html', user = user.get_user_dict(email)))
-    resp.set_cookie("SessionID", sessionID, 60*60*24*365)
-    resp.set_cookie("Email", email, 60*60*24*365)
-    resp.status_code = 200
-    return resp
-
-@dashboard_bp.route("/logout", methods=['POST'])
-def logout():
-    email = request.args.get('email')
-    if email == None or email == "":
-        return "No email provided."
-    resp = make_response(render_template('settings.html'))
-    resp.delete_cookie('SessionID')
-    resp.delete_cookie('Email')
-    return resp
-
-@dashboard_bp.route("/loginrequest", methods=['POST'])
-def loginRequest():
-    email = request.args.get('email')
-    if email is None or email == "":
-        return make_response("No email provided.", 400)
-    
-    result = user.login_request(email)
-    
-    if result[1] != 200:
-        return make_response(result[0], result[1])
-
-    if result[0] is str:
-        return make_response(result[0], result[1])
-    
-    return setCookies(result[0][0], result[0][1])
-
-@dashboard_bp.route("/verify_login")
-def verifyLogin():
-    email = request.args.get('email')
-    if email is None or email == "":
-        return make_response("No email provided.", 400)
-    code = request.args.get('code')
-    if code is None or code == "":
-        return make_response("No code provided.", 400)
-
-    result = user.check_code(email, code)
-    if result[0] == False:
-        return make_response(result[1], 500)
-    
-    return setCookies(email, result[1])
-
-@dashboard_bp.route("/get_user_level", methods=['POST'])
-def getUserLevel():
-    sessionID = request.args.get('SessionID')
-    email = request.args.get('email')
-    if email == None or email == "" or sessionID == None or sessionID == "":
-        return "Couldn't get user level or session"
-    return make_response(user.get_user_level(email, sessionID))
-
-@dashboard_bp.route("/admin/set_user_level", methods=['POST'])
-@required_user_level("USER_LEVEL_ADMIN")
-def setUserLevel():
-    data = request.get_json()
-    if not data:
-        return make_response("No JSON data received", 400)
-    email = data.get('email')
-    level = data.get('level')    
-    if email == None or level == None:
-        return make_response("No email or level specified", 400)
-    
-    userChange = user.get_user_dict(email)
-    
-    if userChange is None:
-        return make_response("Couldn't load user", 500)
-    
-    userChange["level"] = level
-    
-    success = user.update_user(userChange)
-    if not success:
-        return make_response("Failed to update user", 500)
-    
-    return make_response("Successfully updated user", 200)
-
-@dashboard_bp.route('/admin/delete_user', methods=['POST'])
-@required_user_level("USER_LEVEL_ADMIN")
-def deleteUser():
-    data = request.get_json()
-    if not data:
-        return make_response("No JSON data received", 400)
-    
-    email = data.get('email')
-    if email == None:
-        return make_response("No email specified", 400)
-    
-    result = user.delete_user(email)
-    if not result:
-        return make_response(f"Couldn't remove user {email}", 500)
-    
-    return make_response(f"Successfully removed user {email}", 200)
-
-@dashboard_bp.route('/admin/list_users')
-@required_user_level("USER_LEVEL_ADMIN")
-def listUsers():
-    return user.list_users()
 
 ###########################################################
 ###               context functionality                 ###
